refactor: extract method
