+++ PokerHandDojo_20190331/HandType.cs
@@ -1,0 +1,9 @@
+namespace PokerHandDojo_20190331
+{
+    public enum HandType
+    {
+        FourOfAKind,
+        FullHouse,
+        HighCard
+    }
+}
