+++ PokerHandDojo_20190331/IHandKindMatcher.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace PokerHandDojo_20190331
+{
+    public interface IHandKindMatcher
+    {
+        bool Match(IEnumerable<Card> cards);
+        HandType HandType { get; }
+    }
+}
